IT WAS A CASTING ISSUE
In summary, bytes should be one byte long, not multiple bytes long.
Before this commit, in bytesToGeneralData, data[i] (the current
byte) was being interpreted as an int. For values where the MSB
was 1, the sign messed up its value when it was added to var and
implicitly cast to a uint16_t. It also messed up my mental state
for the past two hours. Anyway, by casting data[i] to a uint8_t
before the left-shift operation and adding it to var, the issue
with sign-extension before converting to a uint16_t is gone.