Inserting data into new table; name/time in url
Inserting data into new table still has minor(ish) bugs.
There is no check that the table has been created/a
response was received from the server, so some early
datasets are sent in requests with a null table name.
This has to be fixed for the final/actual implementation.