Chill out when resending
Added a short sleep during resend to allow the server to receive resent
packets in order. Without the sleep, the server received the first
packet generated after connection was regained in the middle of the
resent packets. The sleep allows packets generated after connection is
regained to be properly queued to be resent after the packets that were
generated with no connection have been sent.